+++ zcash_proofs/src/constants.rs
@@ -4,7 +4,9 @@ use bls12_381::Scalar;
 use group::{ff::Field, Curve, Group};
 use jubjub::ExtendedPoint;
 use lazy_static::lazy_static;
-use zcash_primitives::constants::{PEDERSEN_HASH_CHUNKS_PER_GENERATOR, PEDERSEN_HASH_GENERATORS};
+use zcash_primitives::sapling::constants::{
+    PEDERSEN_HASH_CHUNKS_PER_GENERATOR, PEDERSEN_HASH_GENERATORS,
+};
 
 /// The `d` constant of the twisted Edwards curve.
 pub(crate) const EDWARDS_D: Scalar = Scalar::from_raw([
@@ -42,22 +44,22 @@ pub type FixedGeneratorOwned = Vec<Vec<(Scalar, Scalar)>>;
 
 lazy_static! {
     pub static ref PROOF_GENERATION_KEY_GENERATOR: FixedGeneratorOwned =
-        generate_circuit_generator(zcash_primitives::constants::PROOF_GENERATION_KEY_GENERATOR);
+        generate_circuit_generator(zcash_primitives::sapling::constants::PROOF_GENERATION_KEY_GENERATOR);
 
     pub static ref NOTE_COMMITMENT_RANDOMNESS_GENERATOR: FixedGeneratorOwned =
-        generate_circuit_generator(zcash_primitives::constants::NOTE_COMMITMENT_RANDOMNESS_GENERATOR);
+        generate_circuit_generator(zcash_primitives::sapling::constants::NOTE_COMMITMENT_RANDOMNESS_GENERATOR);
 
     pub static ref NULLIFIER_POSITION_GENERATOR: FixedGeneratorOwned =
-        generate_circuit_generator(zcash_primitives::constants::NULLIFIER_POSITION_GENERATOR);
+        generate_circuit_generator(zcash_primitives::sapling::constants::NULLIFIER_POSITION_GENERATOR);
 
     pub static ref VALUE_COMMITMENT_VALUE_GENERATOR: FixedGeneratorOwned =
-        generate_circuit_generator(zcash_primitives::constants::VALUE_COMMITMENT_VALUE_GENERATOR);
+        generate_circuit_generator(zcash_primitives::sapling::constants::VALUE_COMMITMENT_VALUE_GENERATOR);
 
     pub static ref VALUE_COMMITMENT_RANDOMNESS_GENERATOR: FixedGeneratorOwned =
-        generate_circuit_generator(zcash_primitives::constants::VALUE_COMMITMENT_RANDOMNESS_GENERATOR);
+        generate_circuit_generator(zcash_primitives::sapling::constants::VALUE_COMMITMENT_RANDOMNESS_GENERATOR);
 
     pub static ref SPENDING_KEY_GENERATOR: FixedGeneratorOwned =
-        generate_circuit_generator(zcash_primitives::constants::SPENDING_KEY_GENERATOR);
+        generate_circuit_generator(zcash_primitives::sapling::constants::SPENDING_KEY_GENERATOR);
 
     /// The pre-computed window tables `[-4, 3, 2, 1, 1, 2, 3, 4]` of different magnitudes
     /// of the Pedersen hash segment generators.
